Coding improvements
temp: modify tf2onnx/version.py

temp: add dtype=tf.float32 in gemm test to make sure alpha and beta are float instead of double

temp: for now, we do not need gemm in the function rewrite_incomplete_type_support_rs6

temp: conintue when a input is not float

temp: Simplify the logic in gemm_rewriter.py

temp: update test

temp: write a function to obtain the attr for gemm

temp: update graph_matcher.py to modify get_op and _to_pattern to make the functions can return default(None)

temp: coding improvements in gemm_rewriter.py and graph_matcher.py

temp: fix writing error for pylint

temp: delete default=None in _to_pattern and delete the function comments

temp: a spell in comment

temp: delete "None" in get function because the return default is None

temp: delete test_Gemm.py

+++ tests/test_backend.py
@@ -2579,13 +2579,14 @@ class BackendTests(Tf2OnnxBackendTestBase):
         a = tf.placeholder(tf.float32, x_val1.shape, name=_TFINPUT)
         b = tf.placeholder(tf.float32, x_val2.shape, name=_TFINPUT1)
         c = tf.placeholder(tf.float32, x_val3.shape, name=_TFINPUT2)
-        alpha = tf.constant(1.0)
-        beta = tf.constant(2.0)
+        alpha = tf.constant(1.0, dtype=tf.float32)
+        beta = tf.constant(2.0, dtype=tf.float32)
         mul1 = tf.multiply(alpha, tf.matmul(a, b))
         mul2 = tf.multiply(beta, c)
         x_ = mul1 + mul2
         _ = tf.identity(x_, name=_TFOUTPUT)
-        self._run_test_case([_OUTPUT], {_INPUT: x_val1, _INPUT1: x_val2, _INPUT2: x_val3}, graph_validator=lambda g: check_op_count(g, "Gemm", 1))
+        self._run_test_case([_OUTPUT], {_INPUT: x_val1, _INPUT1: x_val2, _INPUT2: x_val3},
+                            graph_validator=lambda g: check_op_count(g, "Gemm", 1))
 
     # test for gemm pattern1: alpha*A*B + C
     def test_gemm_pattern1(self):
@@ -2599,10 +2600,11 @@ class BackendTests(Tf2OnnxBackendTestBase):
         a = tf.placeholder(tf.float32, x_val1.shape, name=_TFINPUT)
         b = tf.placeholder(tf.float32, x_val2.shape, name=_TFINPUT1)
         c = tf.placeholder(tf.float32, x_val3.shape, name=_TFINPUT2)
-        alpha = tf.constant(1.0)
+        alpha = tf.constant(1.0, dtype=tf.float32)
         x_ = tf.multiply(alpha, tf.matmul(a, b)) + c
         _ = tf.identity(x_, name=_TFOUTPUT)
-        self._run_test_case([_OUTPUT], {_INPUT: x_val1, _INPUT1: x_val2, _INPUT2: x_val3}, graph_validator=lambda g: check_op_count(g, "Gemm", 1))
+        self._run_test_case([_OUTPUT], {_INPUT: x_val1, _INPUT1: x_val2, _INPUT2: x_val3},
+                            graph_validator=lambda g: check_op_count(g, "Gemm", 1))
 
     # test for gemm pattern2: A*B + beta*C
     def test_gemm_pattern2(self):
@@ -2616,10 +2618,11 @@ class BackendTests(Tf2OnnxBackendTestBase):
         a = tf.placeholder(tf.float32, x_val1.shape, name=_TFINPUT)
         b = tf.placeholder(tf.float32, x_val2.shape, name=_TFINPUT1)
         c = tf.placeholder(tf.float32, x_val3.shape, name=_TFINPUT2)
-        beta = tf.constant(2.0)
+        beta = tf.constant(2.0, dtype=tf.float32)
         x_ = tf.matmul(a, b) + tf.multiply(beta, c)
         _ = tf.identity(x_, name=_TFOUTPUT)
-        self._run_test_case([_OUTPUT], {_INPUT: x_val1, _INPUT1: x_val2, _INPUT2: x_val3}, graph_validator=lambda g: check_op_count(g, "Gemm", 1))
+        self._run_test_case([_OUTPUT], {_INPUT: x_val1, _INPUT1: x_val2, _INPUT2: x_val3},
+                            graph_validator=lambda g: check_op_count(g, "Gemm", 1))
 
     # test for gemm pattern3: A*B + C
     def test_gemm_pattern3(self):
@@ -2635,7 +2638,8 @@ class BackendTests(Tf2OnnxBackendTestBase):
         c = tf.placeholder(tf.float32, x_val3.shape, name=_TFINPUT2)
         x_ = tf.matmul(a, b) + c
         _ = tf.identity(x_, name=_TFOUTPUT)
-        self._run_test_case([_OUTPUT], {_INPUT: x_val1, _INPUT1: x_val2, _INPUT2: x_val3}, graph_validator=lambda g: check_op_count(g, "Gemm", 1))
+        self._run_test_case([_OUTPUT], {_INPUT: x_val1, _INPUT1: x_val2, _INPUT2: x_val3},
+                            graph_validator=lambda g: check_op_count(g, "Gemm", 1))
 
     def test_graph_matcher(self):
         shape = [2, 6]
